+++ energi/consensus/difficultysim/params/params.go
@@ -20,8 +20,22 @@ const (
 	// AveragingWindow average block window
 	AveragingWindow uint64 = 60 // 60 blocks
 	TargetPeriodGap uint64 = AveragingWindow * TargetBlockGap
+
 	// AsgardIsActive determines the difficulty algorithm that is used for minig
 	AsgardIsActive bool = true
+
+	// BalanaPOSIsActive changes nonce selection algorithm
+	BananaPOSIsActive bool = true
+	BananaStakeCheckDepth uint64 = 0
+
+	// hardfork that activates 15s block time
+	BananaBlockTimeIsActive bool = true
+	BananaTargetBlockGap uint64 = 20
+	BananaMinBlockGap uint64 = 10
+
+	// upon activating BananaDifficultyAdjustment hardfork new gain value will be used
+	BananaDifficultyAdjustment bool = true
+	GainBanana int64 = 15000
 
 	// Parameters for difficulty calculations
 	Gain           int64 = 50000