+++ energi/consensus/posv2_test.go
@@ -35,9 +35,13 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
+const (
+	targetBlockTime = 60
+)
+
 func TestCalculateBlockTimeEMA(t *testing.T) {
 	t.Parallel()
-	emaCalculated := CalculateBlockTimeEMA(testDataBlockTimes, energi_params.BlockTimeEMAPeriod)
+	emaCalculated := CalculateBlockTimeEMA(testDataBlockTimes, energi_params.BlockTimeEMAPeriod, targetBlockTime)
 
 	// check a known value
 	emaExpected58 := uint64(59808819)
@@ -57,7 +61,7 @@ func TestCalculateBlockTimeEMA(t *testing.T) {
 
 func TestCalculateBlockTimeDrift(t *testing.T) {
 	t.Parallel()
-	blockDrift := CalculateBlockTimeDrift(testDataBlockTimeEMA)
+	blockDrift := CalculateBlockTimeDrift(testDataBlockTimeEMA, targetBlockTime)
 
 	// check a known value
 	blockDriftExpected58 := int64(191181)
@@ -519,7 +523,7 @@ func TestPoSDiffV2(t *testing.T) {
 	}
 	// look through tests and assert result
 	for i, tc := range tests {
-		res := CalcPoSDifficultyV2(0, &types.Header{Difficulty: big.NewInt(tc.Parent)}, &TimeTarget{Drift: tc.Drift, Integral: tc.Integral, Derivative: tc.Derivative})
+		res := CalcPoSDifficultyV2(0, &types.Header{Difficulty: big.NewInt(tc.Parent)}, &TimeTarget{Drift: tc.Drift, Integral: tc.Integral, Derivative: tc.Derivative}, true)
 		assert.Equal(t, tc.Result, res.Uint64(), "TC %v", i)
 	}
 }