+++ miner/worker.go
@@ -34,6 +34,7 @@ import (
 	"github.com/energicryptocurrency/energi/event"
 	"github.com/energicryptocurrency/energi/log"
 	"github.com/energicryptocurrency/energi/params"
+	"github.com/energicryptocurrency/energi/energi/api/hfcache"
 	mapset "github.com/deckarep/golang-set"
 
 	energi_params "github.com/energicryptocurrency/energi/energi/params"
@@ -229,13 +230,20 @@ func newWorker(config *params.ChainConfig, engine consensus.Engine, eth Backend,
 
 	// Make sure pending state can be returned for API usage
 	parent := worker.chain.CurrentBlock()
+
+	// get minBlockGap based on banana hardfork activation
+	minBlockGap := energi_params.MinBlockGap
+	if hfcache.IsHardforkActive("Banana-blocktime", parent.NumberU64()) {
+		minBlockGap = energi_params.MinBlockGapBanana
+	}
+
 	num := parent.Number()
 	header := &types.Header{
 		ParentHash: parent.Hash(),
 		Number:     num.Add(num, common.Big1),
 		GasLimit:   core.CalcGasLimit(parent, worker.gasFloor, worker.gasCeil),
 		Extra:      worker.extra,
-		Time:       uint64(parent.Time() + energi_params.MinBlockGap),
+		Time:       uint64(parent.Time() + minBlockGap),
 	}
 	if err := worker.makeCurrent(parent, header); err != nil {
 		panic(err)
@@ -888,13 +896,19 @@ func (w *worker) commitNewWork(interrupt *int32, noempty bool, timestamp int64) 
 	tstart := time.Now()
 	parent := w.chain.CurrentBlock()
 
+	// get minBlockGap based on banana hardfork activation
+	minBlockGap := energi_params.MinBlockGap
+	if hfcache.IsHardforkActive("Banana-blocktime", parent.NumberU64()) {
+		minBlockGap = energi_params.MinBlockGapBanana
+	}
+
 	num := parent.Number()
 	header := &types.Header{
 		ParentHash: parent.Hash(),
 		Number:     num.Add(num, common.Big1),
 		GasLimit:   core.CalcGasLimit(parent, w.gasFloor, w.gasCeil),
 		Extra:      w.extra,
-		Time:       uint64(parent.Time() + energi_params.MinBlockGap),
+		Time:       uint64(parent.Time() + minBlockGap),
 	}
 	// Only set the coinbase if our consensus engine is running (avoid spurious block rewards)
 	if w.isRunning() {